!249 meta2local链接打通 * 解决合入冲突 * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * cleancode * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * 本地申请到空间，不使用rpc备份meta * net_engine 测试通过 * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * update 元数据&接口整改 * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * update 元数据&接口整改 * 检视修改 * rank选择修改 * bm destory修改 * ut修改 * fix alloc bug * Merge remote-tracking branch 'refs/remotes/lfz/br_A3_shm_bm_630_lfz' i… * Merge remote-tracking branch 'refs/remotes/lfz/br_A3_shm_bm_630_lfz' i… * fix compile bug * Merge branch 'br_A3_shm_bm_630' of https://gitee.com/ascend/memfabric_… * fix * fix * fix * bm init修改 * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * fix compile bug * fix * fix * 使用buffer * fix merge bug * Merge remote-tracking branch 'refs/remotes/lfz/br_A3_shm_bm_630_lfz' i… * 修改命名格式 * fix * fix * Merge branch 'br_A3_shm_bm_630' of https://gitee.com/ascend/memfabric_… * fix * fix * fix * bm_proxy实现 * fix ut * fix * fix * bm init * service interface * service interface * service interface * service interface * service 开发 * service 开发 * Merge remote-tracking branch 'refs/remotes/origin/br_A3_shm_bm_630' in… * rpc call alloc success * rpc call alloc success

+++ src/memcache/csrc/entities/mmc_mem_obj_meta.h
@@ -67,7 +67,7 @@ public:
      * @brief Get the priority
      * @return priority
      */
-    uint16_t Priority();
+    uint8_t Priority();
 
     /**
      * @brief Get the number of blobs
@@ -130,7 +130,7 @@ inline uint16_t MmcMemObjMeta::Prot()
     return prot_;
 }
 
-inline uint16_t MmcMemObjMeta::Priority()
+inline uint8_t MmcMemObjMeta::Priority()
 {
     std::lock_guard<Spinlock> guard(spinlock_);
     return priority_;
